# test(ssr): add ssr benchmarks