--- a/packages/@lwc/perf-benchmarks/scripts/fix-deps.sh
+++ b/packages/@lwc/perf-benchmarks/scripts/fix-deps.sh
@@ -14,7 +14,7 @@ set -e
 
 mkdir -p ./node_modules/@lwc
 
-for pkg in @lwc/engine-dom @lwc/engine-server @lwc/synthetic-shadow @lwc/perf-benchmarks-components; do
+for pkg in @lwc/engine-dom @lwc/engine-server @lwc/perf-benchmarks-components @lwc/ssr-runtime @lwc/synthetic-shadow; do
   if [ ! -L "./node_modules/$pkg" ]; then
     ln -s "../../../../../packages/$pkg" "./node_modules/$pkg"
   fi