## Details

~~This will fail until #4702 is merged, but I wanted to get the PR ready.~~ (_Edit: merged!_)

This adds SSR Tachometer/Best benchmarks. For now I'm merely reusing the existing `engine-server` ones so we can compare them.

Interestingly we have some work to do! Running one iteration only (so not statistically significant), I got:

- `tablecmp-render-10k`: 70ms -> 101ms
- `table-render-10k`: 210ms -> 295ms

Probably this is mostly due to the `asyncYield`, but that's fine. We will be able to hopefully see the improvement later when we switch to a no-yield variant. Plus there are probably other optimizations to do.

**Edit:** yep it is due to `asyncYield`! `sync` is way faster:

- `tablecmp-render-10k`: 19ms
- `table-render-10k`: 47ms

## Does this pull request introduce a breaking change?

<!--
    Any change that can cause downstream consumers to fail qualifies as a breaking change.

    Examples:
        - Removing the code for a deprecated API.
        - Adding a new restriction to the compiler which might result in a compilation failure for existing code.
        - Changing the return type of a function in a non-backward compatible fashion.

    Remove the incorrect item for the list.
-->

-   😮‍💨 No, it does not introduce a breaking change.

<!-- If yes, please describe the impact and migration path for existing applications. -->

## Does this pull request introduce an observable change?

<!--
    Observable changes are internal changes that can be observed by downstream consumers.
    Such changes don't qualify as breaking changes because they don't impact any publicly defined
    APIs.

    Examples:
        - Fixing a bug.
        - Changing the invocation timing of a callback, for a callback that has no invocation timing
          guarantee.

    Remove the incorrect item from the list.
-->

-   🤞 No, it does not introduce an observable change.