@nolanlawson 2024-10-28
> the numbers should roughly correlate to a warm-JIT scenario

These benchmarks run 10k renders of the same component with no warmup runs. So it is effectively warm JIT just due to the sheer numbers.

I was thinking of adding tests for 10k different styled components (we have dom benchmarks for these), which would be closer to cold JIT although not quite since the framework code itself (`@lwc/ssr-runtime`) would get invoked plenty of times.

IME our existing Tachometer/Best tests are only good at measuring warm-JIT anyway. For cold JIT we'd have to do ~1 iteration of a huge heterogeneous component tree or something.