@nolanlawson 2024-10-28
Summary of perf numbers (milliseconds):

| Test | Baseline | SSR v2 without transform | SSR v2 with async-yield transform |
| --- | --- | --- | --- |
| `tablecmp-render-10k` | 70 | 101 | 19 |
| `table-render-10k` | 210 | 295 | 47 |